Make combined import export folder, move constants to shared file
