+++ mlte/store/import_export/constants.py
@@ -1,0 +1,19 @@
+"""Constant import export variables."""
+
+MODELS_KEY = "models"
+"""Key for model data in export."""
+
+CUSTOM_LISTS_KEY = "custom_lists"
+"""Key for custom list data in export."""
+
+USERS_KEY = "users"
+"""Key for user data in export."""
+
+CATALOG_KEY = "catalogs"
+"""Key for catalog data in export."""
+
+EXPORT_ZIP_FILE = "store_export.zip"
+"""File name for zipped export file."""
+
+EXPORT_JSON_FILE = "store_export.json"
+"""File name for json export file."""
