+++ mlte/store/import_export/export.py
@@ -15,13 +15,14 @@ from mlte.store.custom_list.store import CustomListStore
 from mlte.store.custom_list.store_session import ManagedCustomListSession
 from mlte.store.user.store import UserStore
 from mlte.store.user.store_session import ManagedUserSession
-
-MODELS_KEY = "models"
-CUSTOM_LISTS_KEY = "custom_lists"
-USERS_KEY = "users"
-CATALOG_KEY = "catalogs"
-EXPORT_ZIP_FILE = "store_export.zip"
-EXPORT_JSON_FILE = "store_export.json"
+from mlte.store.import_export.constants import (
+    CATALOG_KEY,
+    CUSTOM_LISTS_KEY,
+    EXPORT_JSON_FILE,
+    EXPORT_ZIP_FILE,
+    MODELS_KEY,
+    USERS_KEY,
+)
 
 
 class ExportSpec:
