+++ test/store/import_export/conftest.py
@@ -8,7 +8,7 @@ import pytest
 from mlte.store.artifact.store import ArtifactStore
 from mlte.store.catalog.catalog_group import CatalogStoreGroup
 from mlte.store.custom_list.store import CustomListStore
-from mlte.store.export.export import ExportSpec, _export
+from mlte.store.import_export.export import ExportSpec, _export
 from mlte.store.user.store import UserStore
 
 ARTIFACT_EXPORT_DATA: dict[str, Any] = {
@@ -72,16 +72,16 @@ def patched_export():
         return CATALOG_EXPORT_DATA
 
     with patch(
-        "mlte.store.export.export._export_artifacts",
+        "mlte.store.import_export.export._export_artifacts",
         side_effect=_export_artifacts_with_fixtures,
     ), patch(
-        "mlte.store.export.export._export_custom_lists",
+        "mlte.store.import_export.export._export_custom_lists",
         side_effect=_export_custom_lists_with_fixtures,
     ), patch(
-        "mlte.store.export.export._export_users",
+        "mlte.store.import_export.export._export_users",
         side_effect=_export_users_with_fixtures,
     ), patch(
-        "mlte.store.export.export._export_catalogs",
+        "mlte.store.import_export.export._export_catalogs",
         side_effect=_export_catalogs_with_fixtures,
     ):
         yield _export
