+++ test/store/import_export/test_export.py
@@ -12,13 +12,7 @@ from mlte.custom_list.custom_list_names import CustomListName
 from mlte.store.artifact.store_session import ManagedArtifactSession
 from mlte.store.base import StoreType
 from mlte.store.constants import LOCAL_CATALOG_STORE_ID, SAMPLE_CATALOG_STORE_ID
-from mlte.store.export.export import (
-    CATALOG_KEY,
-    CUSTOM_LISTS_KEY,
-    EXPORT_JSON_FILE,
-    EXPORT_ZIP_FILE,
-    MODELS_KEY,
-    USERS_KEY,
+from mlte.store.import_export.export import (
     ExportSpec,
     _export,
     _export_artifacts,
@@ -27,13 +21,21 @@ from mlte.store.export.export import (
     _export_users,
     export_to_file,
 )
+from mlte.store.import_export.constants import (
+    CATALOG_KEY,
+    CUSTOM_LISTS_KEY,
+    EXPORT_JSON_FILE,
+    EXPORT_ZIP_FILE,
+    MODELS_KEY,
+    USERS_KEY,
+)
 from mlte.store.unified_store import UnifiedStore
 from mlte.store.user.policy import user_policy
 from mlte.store.user.store_session import ManagedUserSession
 from mlte.user.model import User
 from test.fixture.artifact import ArtifactModelFactory
 from test.store.conftest import create_test_unified_store
-from test.store.export.conftest import (
+from test.store.import_export.conftest import (
     ARTIFACT_EXPORT_DATA,
     CATALOG_EXPORT_DATA,
     CUSTOM_LIST_EXPORT_DATA,
